+++ src/create-cached-async-fnc.ts
@@ -1,3 +1,4 @@
+import { cloneDeep } from "./clone-deep";
 import { generateId } from "./generate-id";
 import { toConsole } from "./to-console";
 
@@ -54,7 +55,7 @@ export function createCachedAsyncFnc<
 
     if (cache.has(id)) {
       response.status = "HIT";
-      response.data = cache.get(id);
+      response.data = cloneDeep(cache.get(id));
     }
 
     // When a active request exists for this request wait until the first request completes.
@@ -71,7 +72,7 @@ export function createCachedAsyncFnc<
       });
 
       response.status = "HIT";
-      response.data = await completed;
+      response.data = cloneDeep(await completed);
     }
 
     // If no active request or cached request exists for this request. Execute it and save the response to cache
@@ -84,7 +85,7 @@ export function createCachedAsyncFnc<
 
       // Save the data to the cache
       cache.set(id, resolvedData);
-      response.data = resolvedData;
+      response.data = cloneDeep(resolvedData);
 
       // Notify all waiting requests
       const defereds = waitingRequests.get(id);